Applies suggested updates to SDK and API
